Refactor/bigdecimal

+++ src/main/java/br/com/edras/picpaysimplificado/dto/transaction/TransactionResponseDTO.java
@@ -3,6 +3,7 @@ package br.com.edras.picpaysimplificado.dto.transaction;
 import br.com.edras.picpaysimplificado.entity.Transaction;
 import br.com.edras.picpaysimplificado.entity.enums.TransactionStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class TransactionResponseDTO {
@@ -12,7 +13,7 @@ public class TransactionResponseDTO {
     private String payerName;
     private Long payeeId;
     private String payeeName;
-    private Double amount;
+    private BigDecimal amount;
     private LocalDateTime timestamp;
     private TransactionStatus transactionStatus;
 
@@ -29,7 +30,7 @@ public class TransactionResponseDTO {
         this.transactionStatus = transaction.getStatus();
     }
 
-    public TransactionResponseDTO(Long transactionId, Long payerId, String payerName, Long payeeId, String payeeName, Double amount, LocalDateTime timestamp, TransactionStatus transactionStatus) {
+    public TransactionResponseDTO(Long transactionId, Long payerId, String payerName, Long payeeId, String payeeName, BigDecimal amount, LocalDateTime timestamp, TransactionStatus transactionStatus) {
         this.transactionId = transactionId;
         this.payerId = payerId;
         this.payerName = payerName;
@@ -80,11 +81,11 @@ public class TransactionResponseDTO {
         this.payeeName = payeeName;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
